test: executor test 추가 #205
